move metadata_Comparison under advanced_search dir rather than operators dir
