+++ test/models/concerns/advanced_search/filtering_test.rb
@@ -12,7 +12,7 @@ module AdvancedSearch
       include AdvancedSearch::Operators::ComparisonOperators
       include AdvancedSearch::Operators::PatternOperators
       include AdvancedSearch::Operators::ExistenceOperators
-      include AdvancedSearch::Operators::MetadataComparison
+      include AdvancedSearch::MetadataComparison
 
       extend ActiveSupport::Concern
