+++ test/models/concerns/advanced_search/operators/comparison_operators_test.rb
@@ -8,7 +8,7 @@ module AdvancedSearch
       # Test class that includes the ComparisonOperators module
       class TestClass
         include AdvancedSearch::Operators::ComparisonOperators
-        include AdvancedSearch::Operators::MetadataComparison
+        include AdvancedSearch::MetadataComparison
       end
 
       def setup
