+++ test/models/concerns/advanced_search/operators/set_operators_test.rb
@@ -8,7 +8,7 @@ module AdvancedSearch
       # Test class that includes the SetOperators module
       class TestClass
         include AdvancedSearch::Operators::SetOperators
-        include AdvancedSearch::Operators::MetadataComparison
+        include AdvancedSearch::MetadataComparison
         include AdvancedSearch::Operators
 
         # Helper method from parent Operators module
